+++ MauiGeoQuiz.About/Platforms/Windows/PlatformClass1.cs
@@ -1,0 +1,6 @@
+﻿namespace MauiGeoQuiz.About;
+
+// All the code in this file is only included on Windows.
+public class PlatformClass1
+{
+}
