+++ MauiGeoQuiz.About/Platforms/iOS/PlatformClass1.cs
@@ -1,0 +1,6 @@
+﻿namespace MauiGeoQuiz.About;
+
+// All the code in this file is only included on iOS.
+public class PlatformClass1
+{
+}
