+++ MauiGeoQuiz.Core/Services/INavigationService.cs
