+++ MauiGeoQuiz.Core/Services/IStringService.cs
