+++ MauiGeoQuiz.Core/UseCases/ICommandUseCase.cs
