+++ MauiGeoQuiz.Core/UseCases/ISyncUseCase.cs
