+++ MauiGeoQuiz.Core/UseCases/IUseCase.cs
