+++ MauiGeoQuiz/AppShell.xaml.cs
@@ -1,7 +1,7 @@
-﻿using MauiGeoQuiz.About;
+﻿using MauiGeoQuiz.Core.Constants;
 using MauiGeoQuiz.Game;
 using MauiGeoQuiz.MainMenu;
-using MauiGeoQuiz.Presentation.Constants;
+using MauiGeoQuiz.ViewPages;
 
 namespace MauiGeoQuiz;
