+++ MauiGeoQuiz/ViewPages/AboutPage.xaml
@@ -3,11 +3,11 @@
     xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
     xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
     xmlns:maui="clr-namespace:ReactiveUI.Maui;assembly=ReactiveUI.Maui"
-    xmlns:vm="clr-namespace:MauiGeoQuiz.Presentation.About;assembly=MauiGeoQuiz.Presentation"
+    xmlns:vm="clr-namespace:MauiGeoQuiz.About.ViewModels;assembly=MauiGeoQuiz.About"
     xmlns:strings="clr-namespace:MauiGeoQuiz.Resources.Strings"
     x:TypeArguments="vm:AboutViewModel"
     x:DataType="vm:AboutViewModel"
-    x:Class="MauiGeoQuiz.About.AboutPage">
+    x:Class="MauiGeoQuiz.ViewPages.AboutPage">
 
     <ScrollView>
